+++ NativeEndpointWorkspace/Properties/AssemblyInfo.cs
@@ -11,4 +11,4 @@ using System.Windows;
 [assembly: ThemeInfo(ResourceDictionaryLocation.None, ResourceDictionaryLocation.SourceAssembly)]
 [assembly: AssemblyVersion("0.0.2.0")]
 [assembly: AssemblyFileVersion("0.0.2.0")]
-[assembly: AssemblyInformationalVersion("0.0.2rc02")]
+[assembly: AssemblyInformationalVersion("0.0.2rc03")]